decrease Prompt's completion memory usage
Storing a set of seen entries just to avoid calling list.remove() is a
bit excessive here, considering that the user is more likely to lookup
an entry close to the beginning of the list and the realistic number of
entries is <100_000.

I've tried to scale this lookup "algorithm" to handle more entries, but
whatever I throw at it it just keep getting worse due to any indirection
in Python being a bottleneck. Sifting through lookups with references to
strings fixed in memory is just too much for this language, the overhead
is unbearable, anything more clever than lists and dicts is causing more
harm than good.

I'm not dismissing the possibility of a more complex solution existing
that is superior to status quo (lower memory usage and runtime cost),
but it's way more unlikely for it to exist in Python than in a natively
compiled language. So, the simplest solution stays...